pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ondepi"
version = "0.1.0"
description = "Headless audio live source streamer for AzuraCast"
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [
  { name = "OndePi" }
]
keywords = ["audio", "streaming", "icecast", "azuracast", "raspberry-pi"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License"
]
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pyserial>=3.5",
  "sounddevice>=0.4.6",
  "numpy>=1.26",
  "tomli>=2.0.1; python_version < '3.11'",
  "tomli-w>=1.0.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0"
]

[project.scripts]
ondepi = "ondepi.main:main"
ondepi-cli = "ondepi.cli:main"
